Add Timers to dropoff and package objects
Created a timer class.  Added timers to the dropoff and package objects.
For now they are all just 10 second timers, but in the future it would
be cool to make them depend on how far away the courier is.

I've noticed that a packages and dropoffs are essentially the same
thing.  I should probably merge them into a single class called
Location/Address.