Add images to courier, redo control code
Added the car images to the courier.

Used all four directions (up, down, left, right) in the Courier code.
Changed how the car is controlled.  Now it is controlled with gas and
direction.  I'm not sure if I like it yet.  Also used the new Controller
class to get interpret user input.