Add sound effects
Added sound effects for picking up a delivery, dropping off a delivery,
crashing, and missing a delivery.  Created sound effects with sfxr.

I committed the .sfs files because that will allow me to make minor changes
to the the sound effects in sfxr, and then I can revert if I don't like the
changes.  However, those files are binary.